requirements.txt
qai-hub
faiss_cpu
numpy
opencv-python-headless
Pillow
python_Levenshtein
qai_hub_models
rapidfuzz
streamlit
torch
transformers
timm
einops
ftfy
